chore(wmg): move ScreenTint to wmgv2
