+++ frontend/src/views/WheresMyGeneV2/components/ScreenTint/style.ts
